+++ examples/optimize_vertical_silicon_coupler.py
@@ -18,7 +18,7 @@ from fdtdx.fdtd.backward import full_backward
 from fdtdx.fdtd.fdtd import reversible_fdtd
 from fdtdx.interfaces.recorder import Recorder
 from fdtdx.objects.boundaries.initialization import BoundaryConfig, boundary_objects_from_config
-from fdtdx.objects.container import ArrayContainer, ParameterContainer
+from fdtdx.fdtd.container import ArrayContainer, ParameterContainer
 from fdtdx.objects.detectors.energy import EnergyDetector
 from fdtdx.objects.detectors.poynting_flux import PoyntingFluxDetector
 from fdtdx.objects.initialization import apply_params, place_objects
