+++ examples/speed_benchmark.py
@@ -23,7 +23,7 @@ from fdtdx.core.plotting import colors
 from fdtdx.fdtd.fdtd import custom_fdtd_forward
 from fdtdx.interfaces.recorder import Recorder
 from fdtdx.objects.boundaries.initialization import BoundaryConfig, boundary_objects_from_config
-from fdtdx.objects.container import ArrayContainer
+from fdtdx.fdtd.container import ArrayContainer
 from fdtdx.objects.detectors.energy import EnergyDetector
 from fdtdx.objects.detectors.poynting_flux import PoyntingFluxDetector
 from fdtdx.objects.initialization import apply_params, place_objects
