+++ src/fdtdx/fdtd/__init__.py
@@ -9,6 +9,7 @@ method for electromagnetic wave propagation, including:
 """
 
 from fdtdx.fdtd.backward import full_backward
+from fdtdx.fdtd.container import ArrayContainer, ParameterContainer, ObjectContainer, SimulationState
 from fdtdx.fdtd.fdtd import (
     checkpointed_fdtd,
     custom_fdtd_forward,
@@ -20,4 +21,8 @@ __all__ = [
     "checkpointed_fdtd",  # FDTD with field state checkpointing
     "custom_fdtd_forward",  # Customizable FDTD forward propagation
     "full_backward",
+    "ArrayContainer",
+    "ParameterContainer",
+    "ObjectContainer",
+    "SimulationState",
 ]
