+++ src/fdtdx/fdtd/backward.py
@@ -7,7 +7,7 @@ import jax
 from fdtdx.config import SimulationConfig
 from fdtdx.fdtd.update import add_interfaces, update_detector_states, update_E_reverse, update_H_reverse
 from fdtdx.objects.boundaries.periodic import PeriodicBoundary
-from fdtdx.objects.container import ObjectContainer, SimulationState
+from fdtdx.fdtd.container import ObjectContainer, SimulationState
 
 
 def cond_fn(state, start_time_step: int) -> bool:
