+++ src/fdtdx/fdtd/fdtd.py
@@ -9,7 +9,7 @@ from fdtdx.fdtd.backward import backward
 from fdtdx.fdtd.forward import forward, forward_single_args_wrapper
 from fdtdx.interfaces.state import RecordingState
 from fdtdx.objects.boundaries.perfectly_matched_layer import BoundaryState
-from fdtdx.objects.container import ArrayContainer, ObjectContainer, SimulationState, reset_array_container
+from fdtdx.fdtd.container import ArrayContainer, ObjectContainer, SimulationState, reset_array_container
 from fdtdx.objects.detectors.detector import DetectorState
