+++ src/fdtdx/fdtd/update.py
@@ -4,7 +4,7 @@ import pytreeclass as tc
 from fdtdx.config import SimulationConfig
 from fdtdx.fdtd.curl import curl_E, curl_H, interpolate_fields
 from fdtdx.objects.boundaries.periodic import PeriodicBoundary
-from fdtdx.objects.container import ArrayContainer, ObjectContainer
+from fdtdx.fdtd.container import ArrayContainer, ObjectContainer
 from fdtdx.objects.detectors.detector import Detector
 from fdtdx.shared.misc import add_boundary_interfaces, collect_boundary_interfaces
