+++ src/fdtdx/shared/misc.py
@@ -3,7 +3,7 @@ from typing import Sequence
 import jax
 
 from fdtdx.objects.boundaries.perfectly_matched_layer import PerfectlyMatchedLayer
-from fdtdx.objects.container import ArrayContainer
+from fdtdx.fdtd.container import ArrayContainer
 
 
 def collect_boundary_interfaces(
